- moved PlayerTasks to a separate file. - JSDocs added
